pyproject.toml
[tool.poetry]
name = "gtk-kugaconstruct"
version = "0.1.0"
description = ""
authors = ["Narashi Kuga <orenvady@gmail.com>"]
readme = "README.md"
packages = [{include = "gtk_kugaconstruct"}]

[tool.poetry.dependencies]
python = "^3.11"
pygobject = "^3.46.0"
psycopg2-binary = "^2.9.9"
faker = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
